(service-crosscheck) rename: trip repo >> crosscheck repo
