+++ api/services/crosscheck/src/interfaces/repository/CrosscheckRepositoryProviderInterface.ts
@@ -2,13 +2,13 @@ import { ParentRepositoryProviderInterface, ParentRepositoryProviderInterfaceRes
 import {Trip} from '../../entities/trip';
 import {PersonInterface} from '../TripInterface';
 
-export interface TripRepositoryProviderInterface extends ParentRepositoryProviderInterface {
+export interface CrosscheckRepositoryProviderInterface extends ParentRepositoryProviderInterface {
   findByOperatorJourneyIdAndOperatorId(params: { operator_journey_id?: string, operator_id: string }): Promise<Trip>;
   findByPhoneAndTimeRange(phone: string, startTimeRange: { min: Date, max: Date }): Promise<Trip>;
   findByIdAndPushPeople(id: string, people: PersonInterface[], newStartDate: Date): Promise<Trip>;
 }
 
-export abstract class TripRepositoryProviderInterfaceResolver extends ParentRepositoryProviderInterfaceResolver {
+export abstract class CrosscheckRepositoryProviderInterfaceResolver extends ParentRepositoryProviderInterfaceResolver {
   public async findByOperatorJourneyIdAndOperatorId(params: { operator_journey_id?: string, operator_id: string }): Promise<Trip> {
     throw new Error();
   }
